Merged PR 147: Remove Private DNS Zone for adf.azure.com
adf.azure.com is used to create a common "portal" A record.  This "portal" A record is used by all ADF deployments, therefore when an ADF instance is deleted, the "portal" A record is also removed which causes DNS resolution to fail for all other ADF deployments.

Since portal.adf.azure.com is a SaaS portal, we will remove private dns zone resolution for this endpoint.

**NSLookup from Spoke VNET:**

![image.png](https://dev.azure.com/CanadaESLZ/631e2ac5-7e1c-4107-a224-3c9445f31024/_apis/git/repositories/31a7fa88-8b30-4537-81c2-ebd936ff144b/pullRequests/147/attachments/image.png)

Related work items: #176